Attempt to unify xunit and nunit step acpects; WIP
