+++ Allure.NUnit.Examples/AllureStepLogTests.cs
@@ -1,4 +1,5 @@
 using System;
+using Allure.Net.Commons.Steps;
 using NUnit.Allure.Attributes;
 using NUnit.Allure.Core;
 using NUnit.Framework;
