+++ Allure.NUnit/Attributes/AllureBeforeAttribute.cs
@@ -1,14 +1,11 @@
-using System;
-using AspectInjector.Broker;
-using NUnit.Allure.Core.Steps;
+using Allure.Net.Commons.Steps;
+using NUnit.Allure.Core;
 
 namespace NUnit.Allure.Attributes
 {
-    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Constructor)]
-    [Injection(typeof(AllureStepAspect))]
-    public class AllureBeforeAttribute : AllureStepBaseAttribute
+    public class AllureBeforeAttribute : AllureStepAttributes.AbstractBeforeAttribute
     {
-        public AllureBeforeAttribute(string name = null) : base(name)
+        public AllureBeforeAttribute(string name = null) : base(name, AllureNUnitHelper.ExceptionTypes)
         {
         }
     }
