+++ Allure.Net.Commons/Allure.Net.Commons.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFrameworks>net45;netstandard2.0</TargetFrameworks>
+        <TargetFramework>netstandard2.0</TargetFramework>
         <Version>2.10-SNAPSHOT</Version>
         <GeneratePackageOnBuild>false</GeneratePackageOnBuild>
         <Authors>Alexander Bakanov, Nikolay Laptev</Authors>
@@ -25,6 +25,10 @@
 
     <ItemGroup>
         <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.1.1" PrivateAssets="All" />
+    </ItemGroup>
+    
+    <ItemGroup>
+        <PackageReference Include="AspectInjector" Version="2.8.1" />
     </ItemGroup>
 
     <ItemGroup>
