+++ Allure.Net.Commons/Steps/AbstractSkipAttribute.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Allure.Net.Commons.Steps
+{
+    [AttributeUsage(AttributeTargets.Parameter)]
+    public abstract class AbstractSkipAttribute : Attribute
+    {
+    }
+}
