+++ Allure.XUnit.Examples/ExampleStepAttributes.cs
@@ -11,8 +11,8 @@ public class ExampleStepAttributes : IDisposable
     [AllureBefore("Initialization in constructor")]
     public ExampleStepAttributes()
     {
-        AddAttachment();
-        NestedStep(1);
+        // AddAttachment();
+        // NestedStep(1);
         NestedStepReturningString("Second");
     }
