+++ Allure.XUnit/Allure.XUnit.csproj
@@ -38,7 +38,7 @@
         <PackageReference Include="AspectInjector" Version="2.8.1" />
         <PackageReference Include="xunit.assert" Version="2.4.1" />
         <PackageReference Include="xunit.core" Version="2.4.1" />
-        <PackageReference Include="xunit.runner.reporters" Version="2.4.2"/>
+        <PackageReference Include="xunit.runner.reporters" Version="2.4.2" />
     </ItemGroup>
 
     <ItemGroup>
