+++ Allure.XUnit/AllureXunitHelper.cs
@@ -3,15 +3,19 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using Allure.Net.Commons;
+using Allure.Net.Commons.Storage;
 using Allure.XUnit;
 using Allure.Xunit.Attributes;
+using Allure.XUnit.Attributes;
 using Xunit;
 using Xunit.Abstractions;
+using Xunit.Sdk;
 
 namespace Allure.Xunit
 {
     public static class AllureXunitHelper
     {
+        internal static List<Type> ExceptionTypes = new List<Type> { typeof(XunitException) };
         static AllureXunitHelper()
         {
             const string allureConfigEnvVariable = "ALLURE_CONFIG";
