+++ Allure.XUnit/Attributes/IAllureInfo.cs
@@ -1,4 +1,4 @@
-namespace Allure.Xunit
+namespace Allure.XUnit.Attributes
 {
     internal interface IAllureInfo
     {
