+++ Allure.XUnit/Attributes/Steps/AllureBeforeAttribute.cs
@@ -1,13 +1,11 @@
-using System;
-using AspectInjector.Broker;
+using Allure.Net.Commons.Steps;
+using Allure.Xunit;
 
 namespace Allure.XUnit.Attributes.Steps
 {
-    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Constructor)]
-    [Injection(typeof(AllureStepAspect))]
-    public class AllureBeforeAttribute : AllureStepBaseAttribute
+    public class AllureBeforeAttribute : AllureStepAttributes.AbstractBeforeAttribute
     {
-        public AllureBeforeAttribute(string name = null) : base(name)
+        public AllureBeforeAttribute(string name = null) : base(name, AllureXunitHelper.ExceptionTypes)
         {
         }
     }
