+++ Allure.XUnit/Attributes/Steps/AllureStepAttribute.cs
@@ -1,11 +1,11 @@
-using AspectInjector.Broker;
+using Allure.Net.Commons.Steps;
+using Allure.Xunit;
 
 namespace Allure.XUnit.Attributes.Steps
 {
-    [Injection(typeof(AllureStepAspect))]
-    public class AllureStepAttribute : AllureStepBaseAttribute
+    public class AllureStepAttribute : AllureStepAttributes.AbstractStepAttribute
     {
-        public AllureStepAttribute(string name = null) : base(name)
+        public AllureStepAttribute(string name = null) : base(name, AllureXunitHelper.ExceptionTypes)
         {
         }
     }
